FIX : TimeZone 오류 수정 및 운동 칼로리 계산 로직 수정
realtime -> 밀리초 적용, Long 타입으로 변경
TimeZone -> Asia/Seoul

+++ API_container/MadPtApi/src/main/java/com/example/MadPtApi/domain/Record.java
@@ -27,7 +27,7 @@ public class Record {
 
     private LocalDateTime endTime;
 
-    private int realTime;
+    private Long realTime;
 
     @ManyToOne(fetch = LAZY)
     @JoinColumn(name = "member_id")
